refactor folder structure, implement redux for theme selection state
