+++ components/Display.js
